Fix prezzo prodotto Carrello
il valore passato in post di un prodotto veniva assegnato a 0, e non al suo reale prezzo, ora il carrello riceve anche i prezzi dei prodotti in modo corretto

+++ Candy/Pages/Index.cshtml.cs
@@ -18,7 +18,7 @@ namespace Candy.Pages
         public void OnGet()
         {
         }
-        public void OnPost(int quantProd, string nomeProd, int prezProd, string btn)
+        public void OnPost(int quantProd, string nomeProd, double prezProd, string btn)
         {
             this.prod.quantita = quantProd;
             this.prod.nome = nomeProd;
